backend/requirements.txt
asgiref==3.11.1
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==42.0.2
Django==4.2.7
django-cors-headers==4.3.1
django-filter==23.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.14
gunicorn==21.2.0
idna==3.11
inflection==0.5.1
packaging==26.0
psycopg2-binary==2.9.9
pycparser==3.0
PyJWT==2.11.0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.3
razorpay==2.0.0
requests==2.31.0
sqlparse==0.4.4
uritemplate==4.2.0
urllib3==2.6.3
whitenoise==6.6.0
